Gemini API/2 SQL Query/requirements.txt
chromadb
faiss-cpu
google-generativeai
langchain_google_genai
langchain_experimental
langchain
langchain-community
mysql-connector-python
protobuf
pymysql
python-dotenv
sentence-transformers
streamlit
tiktoken